[mod] moved more menus
