+++ src/components/main/MainImageGrid/MainImageGrid.tsx
@@ -6,7 +6,7 @@ import { Container, Box, Tabs, Tab, Grid } from "@mui/material";
 import { useDndFileDrop, useContextMenu, useHotkeys } from "hooks";
 
 import { MainImageGridItem } from "../MainImageGridItem";
-import { ImageCategoryMenu } from "../ImageCategoryMenu";
+import { ImageCategoryMenu } from "../../menus/ImageMenu/ImageCategoryMenu";
 import { MainImageGridAppBar } from "../MainImageGridAppBar";
 
 import {
