+++ src/components/main/MainImageGridAppBar/MainImageGridAppBar.tsx
@@ -26,7 +26,7 @@ import {
 import { useDialogHotkey } from "hooks";
 
 import { DialogWithAction } from "components/dialogs";
-import { ImageCategoryMenu } from "../ImageCategoryMenu";
+import { ImageCategoryMenu } from "../../menus/ImageMenu/ImageCategoryMenu";
 import { TooltipTitle } from "components/common/tooltips";
 
 import {
